Simulate brush pressure based on distance between move events.
This is inherently affected by how busy the CPU is. A busy cpu skips
move events and you end up with lighter lines than expected.

The touch device pressure is already hacked to twice the pressure of
mouse events, based on experimentation with the android emulator.

Fast mobiles will have different results to slow mobiles.

Ideally the pressure can be calibrated, or even better, real pressure
could be used in devices with that info.

At least this is a cool start and looks reasonable.